BLE Host - Add comments; move decls to priv hdrs.
* Add API documentation in comments above some public functions.
* Move prototypes for non-API functions into "[...]_priv.h" headers.
* Delete some obsolete and unused functions.
